Moved inventory hover component to station components
