+++ source/core/src/main/com/csse3200/game/entities/factories/StationFactory.java
@@ -9,9 +9,8 @@ import com.badlogic.gdx.physics.box2d.BodyDef.BodyType;
 import com.csse3200.game.components.FlameComponent;
 import com.csse3200.game.components.TooltipsDisplay;
 import com.csse3200.game.components.player.InventoryComponent;
+import com.csse3200.game.components.station.InventoryDisplayHoverComponent;
 import com.csse3200.game.components.station.*;
-import com.csse3200.game.components.player.InventoryDisplayHoverComponent;
-import com.csse3200.game.components.station.StationProgressDisplay;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.physics.PhysicsLayer;
 import com.csse3200.game.physics.PhysicsUtils;
